Move files into floodfilling widget:
Many of the necessary functionality for running floodfilling were made
in CATMAID, since some functionality could be useful for other
applications. However at the moment it is easier to keep things seperate
especially for keeping up with changes being made upstream. Keeping
things seperate removes a lot of the work of refactoring migrations and
reducing conflicts on merges