Job visualization:
fleshed out job tables to now show more information. Added search
functionality for columns.

cloned the OptionsDisplay widget for adaptation into a results popup
window. The results returned by floodfilling include huge volumes that
will probably be slow to render, thus this popup will be used to show
summary statistics to determine if the run should be investigated
further or if it can be scrapped. The popup will also be useful for
selecting specific regions for further investigation.