Improve layer refreshing
If you click on a new result, it would be nice if the layers would
automatically display the correct new information.

Currently the segmentation layer seems to be cached somewhere. The layer
is refreshed correctly and checking the get requests shows that the
correct n5 file is being queried, but it still displays old segmentations